updated trainings + ui
